fixes bug26 (edmac fails if modulename = macroname)
edmac foo failed if foo is both a macro name and a macrolib name.
See [#26]. The problem was in MacroServer.getMacroInfoObj('foo'), which
returned None if it found a macrolib with name foo without checking if
a macro named 'foo' existed as well. Now we use of
"getElementWithInterface" API to do it well.

Fixes: [#26]