Use getModelType instead of getModelClass in type checks
Previously, to check the model type (e.g. attribute, or device) the
code in TaurusValue relied on comparing the model class against
TaurusAttribute. This is not safe because the class can be a subclass
(e.g. TangoAttribute). Now TaurusElementType is used instead of the
model class. Also, in order to be able to call getTaurusElementType
from classes and not from instances, we convert getTaurusElementType
methods into classmethods for taurusAttribute, TaurusDevice, etc.