Fix bug-44. Change verbosity of log messages
Bug-44 reported an exception, but it was in fact a handled exception
that was simply being reported too verbosely via the log system. This
commit changes the verbosity of this and other similar messages from
WARNING to DEBUG. It also fixes (or marks with comments) a few issues
found by PyLint in the same files.