Fix Bug-51. Get modulename when creating a PanelDescription from a panel
PanelDescription.fromPanel() was not setting the module name when
creating PanelDescription object from an existing Panel. This is ok for
classes known to the TaurusWidgetFactory, but not in general. This fix
keeps the same behaviour for the Taurus classes but stores the module
name for the rest of the classes.