Fix bug-55: Create bat wrappers for scripts on windows installs
Windows does not recognise extensionless scripts. We create bat
wrappers to allow windows users to call the scripts in a similar
way as the posix users. The solution is based on this recipe:
http://matthew-brett.github.io/pydagogue/installing_scripts.html
but we give the user control over whether the python interpreter
found in the shebang of the script should be honored by the
wrappers or not (via the new --wrapper and --ignore-shebang
options of the install_scripts distutils command).
We now permit building under windows even if rcc is not
available (skipping rcc creation if --with-tango-icons is used)

This fixes Bug-55