Add compact property menu option to TaurusForm
TaurusForm context menu now has an action to set the compact mode
for all the items in the form. Also, this property is stored in
configuration, so it is remembered in perspectives of taurusguis.
Finally, the setCompact methid now forces a refresh of all items,
so the change is applied inmediately