Allow to set custom TextInteractionFlag
Define a _dynamicTextInteractionFlags which is enabled by
default and disabled when we use setTextInteractionFlags.
Overwrite the QLabel property 'textInteractionFlags' and
allow to customize it with QtDesigner. The method
setDynamicTextInteractionFlags is used to request a new
dynamic flag.

Note from committer: I had to modify the original patch
by Valentin Valls in order for it to apply cleanly. I
also corrected some comments and reformatted the commit
message