Reduce log level of TangoAttribute.pushEvent
Since a commit (d204d9) TangoAttribute.push_event logs an error whenever
a polling needs to be activated, but does it even if polling is already
active, causing a flood of error messages. Fix it by a) reducing the
level to info (being a non-critical issue) and b) avoiding logging if
polling is already active.