Support more JDraw properties for text item in the JDraw synoptic
It feature a new class QGraphicsTextBoxing to contains information from
JDraw text item (origin, shape, colors...) This class manage the
rendering by composition (it use a text and a rect item). The parsing code
was improved to support this new configurations. The alignment is done by
computing a location for the children text item when something change.

* Provide a basic support of text alignment (text alignment for multi-lines is not yet done)
* Provide background/foreground color
* Move the internal scaling of text (0.8) in the text sub-component
* Move the parsing of font property in getLabelObj (this code was not reachable when this method was not called, no?)