Set default stop_delay to 0.1
stop_delay is None by default, which means "infinite". This is not the
correct default (it would always fail). Change default stop_delay to 0.1

Eventually it may be a better idea to change meaning of stop_delay=None
to mean stop_delay=0 or else use 0 as default, but currently
stop_delay=0 fails. Hence the 0.1

Also, simplify dynamic doc setting

refs: #4387