Fix bug-168: build docs works with test modules
Testing modules: test_scan and test_wm try to retrieve handy motor names
from the sar_demo environment. This is not possible if the door
does not exist. The documentation build should not require any door.
The exception is captured and the corresponding warning message logged.
This way the documentation build works.