Fix bug-169: exception when using parent model
For the Tango scheme (and possibly others), there is an inconsistency
between the return value of TaurusBaseWidget.getModelClass for
Configuration models: the return value depends on whether parent model
feature is used or not. In case of not using the parent model,
getModelClass returns the appropriate configuration class. In contrast,
if using the parent model (with parent-->device and
child-->attr?configuration) getModelClass returns the configuration
*proxy* class. This last case is the root cause of bug-169.

Make TaurusBaseWidget.getModelClass return the configuration class and
not the proxy class in all situations