Fix typo in config property key for TaurusValue "labelConfig"
TaurusValue defines a "labeConfig" key for a configuration property.
Change it to "labelConfig".
Note: this has the undesired effect of not loading info in previously
stored settings files which used "labeConfig" and thus is a potential
(minor/aesthetic) loss of user configuration, but it is justified
since this is a relatively new feature (not widely used yet) and it
can easily be reconfigured at run time by the same user.