Fix bug-195: allow setting a TaurusCommandButton without text
Setting thee CustomText property of TaurusCommandButton to None has the
same effect than setting it to an empty string. In both cases it means
"use the command name as text". Change this so that setting it to an
empty string means "use an empy string" (i.e., "show no text").