Stop ignoring exclude_patterns
Auto_rst4API_Creator.documentModule admits a exclude_patterns keyword
argument, but it is ignored in favour of the exclude_patterns given in
the constructor. Fix this by using the patterns given in the kwarg and
reverting to the ones given in the constructor only if the kwarg is not
passed.