Modify doc of SendToCtrl
SendToCtrl allows to controll the hardware in a non standard (API) way.
It accepts any string, and can return any string, and the executed logic
is up to the controller code.

In case of execution of Pool's SendToController command (Tango), and no return value of the
controller's SendToCtrl method the following error is raised:
"omniORB: Assertion failed.  This indicates a bug in the application"

Document the obligatory return value.