Propagate taurusValueBuddy ref to subwidgets in compact mode
When a TaurusValue updates its ReadWidget and write Widgets, it inserts
a weak ref to itself (called taurusValueBuddy) into the new ReadWidget
and writeWidget. When it goes into compact mode, the ReadWidget becomes
a Switcher that contains both a a read and write subwidget (and the
WriteWidget is removed). These subwidgets of the switcher may also
require a reference to the "taurusValueBuddy" (for example in the
case of complex TaurusValue-based widgets such as the PoolMotorTV).
Insert the weakref also into these subwidgets of the Switcher.

a Switcher. a switcher is used for
its ReadWidget.