Fix bug-242: inconsistent pending ops in TaurusValueCombobox
In some cases, when one creates a TaurusValueCombobox and sets the
model after having called addValueNames, the widget is left in an
inconsistent "pending operations" state (the write value shown is
the same as the one applied, but the widget reports having pending
operations -i.e. it is "blue").
Fix it by calling emitValueChanged() at the end of the setModel
method.