Use SpockInputHandler and temp. disable input hook
This is a proof-of-concept of the CL mode used to handle interactive
macros in spock.

Use SpockInputHandler instead of InputHandler (raw_input instead of
Qt dialogs).

Temporarily disable the PyOS_InputHook, in order to avoid problems with some
ipython versions in the qt gui mode. It will have a side effect: QtEventLoop
will not be executed when the execution thread is stopped in the raw_input.
This can affect the spock GUI elements, in the current state: plots and showscan.
More details in: https://sourceforge.net/p/sardana/tickets/253/#9c7f