Implement feat-308: Add filter for same-value events
Add RepeatedEventFilter class filter to taurus.core.util.eventfilters.
Its instances keep track of the last event for each source and type and blocks
any periodic or change events whose evt_value.value is identical to the
previous one. If evt_value.value is not available, the whole evt_value is used
for comparison and as a future reference.