Make some installed files executable
When installing with setup.py install, the files from packages and
package_data are installed with the standard mask permissions,
regardless of their initial permisions. This prevents resource files
such as the Timeout script to be used (in this case it produces an 
error in a test) from installed files. Fix it by making setup.py 
install to set executable permissions to any file listed in the
executable_data variable.