Install ui dirs as package_data instead of packages
"ui" dirs are treated by setup.py as packages (submodules) and
installed as such. This forces the ui dir to have a __ini__.py file.
This was necessary when ui dirs also contained compiled ui_*.py files,
but since SEP11 it is no longer the case. Therefore, remove all
__init__.py files from ui dirs and install the ui dirs as package_data
instead of packages. Also force the ui files to be added explicitly.