Allow to specify a static source of thumbnails
Modify setup.py build_doc command to use --thumbnails-source argument
instead of --external-img-tools. The new argument accepts a source dir
or zipfile of pre-created thumbnails. This is required for RTD because
it cannot PyQt or wand.

Also add a pre-created thumbnails.zip file. Note that this file needs
to be updated manually (build_doc generates a ~thumbnails.zip file
which can be copied)