Do not import taurusdevicetree in tree module
The taurusdevicetree module is not being used anywhere in Taurus and 
depends on non-standard and non-provided modules. It is also quite
specific and tango-centric, and therefore difficult to maintain, so it
should be removed from Taurus or merged with Taurusdbtree. 
As a first step towards this, do not import it in the tree module (so
we do not need to maintain it but we facilitate the transition for 3rd
party code using it -e.g. Vacca-)