Deprecate OperationMode logic in taurus.core
taurus Manager and Factory define an inconsistent logic to set
the mode which taurus should work (Online or Offline).

The OperationMode is used only for the taurusdesigner logic
and most of the part of the code is inconsistent.

This commit moves to deprecated all the logic and removes the calls
in tarurus where it is used.