Remove simulator scheme
Simulator scheme is a mock for the tango scheme when the
taurus.core.OperationMode is offline. It  was used by taurusdesigner.

This logic has been deprecated in the TEP3 and according with it
the protection for the designer should be defined in each widget
implementation and not in the taurus.core.