Add feat-161: maintain sort order of sets
When exporting data from a plot (or trend) to ascii, the sorting
order is alphabetical, which may not correspond to the order in
which the curves are ordered in the plot, and this confuses the
users. Use Z-order of the plot instead of alphabetical.

Also, fix a bug preventing the export of "all sets in one file"
feature.