Move bck-compat code from TaurusAttrValue to the schemes
TaurusAttrValue provides some API for bck-compat which should
belong to the {Tango,Evaluation}AttrValue classes. Move it.
Also improve the __getattr__ exception handling in those classes

Note: this commit leaves the TangoAttribute in a broken state:
.read() returns none unless cache=False is used.
Possible workaround is to set cache=False in read() if
self.__atts_value is None.
Investigate cause and best solution