Fix bug #159 TaurusWheelEdit writes 0
TaurusWheelEdit writes 0 to attribute when the w_value is
higher than the TaurusWheelEdit limits.
Fix it, setting the ranges in the consturctor to  +inf, -inf by default
instead of the max/min possible value acording with the current format.

The limits can be defined using the proper methods or defining the limits
in the attribute configuration (e.g. using jive for a tango attribute)