Fix bug-167: Show context menu of TaurusImageDialog
When a widget TaurusImageDialog is docked in the Main Window, the
shown context menu is the one of the Main Window and it is not
the one of the TaurusImageDialog.

Show the context menu of TaurusImageDialog when right-clicking in
a TaurusImageDialog widget, even if this widget is docked in the
Main Window.

Solve this bug also for TaurusCurveDialog and for TaurusTrendDialog.