Fix memleak in TaurusPlot (partial fix of bug-171)
Qwt has a mem leak in the QwtScaleEngine.transformation() method,
which is used in several parts of Taurus (the TaurusPlot.setData
method among them) to obtain the transformation type of a given
axis(i.e., linear or log scale).
Minimize the impact of this Qwt bug by caching the transformation
type and providing a new getter for the transformation type.