Change width of default pen (workaround for bug-171)
TaurusPlot defines a set of pens to use by default. They are set
with width=2. It has been found that using width=2 causes a
memory leak on replotting. Change the width to 1 to avoid
triggering this mem leak. Note that this is just a workaround for
the *default* pens. The leak will still appear if the pen is
customised (e.g. by the user)