Implement feat-175: Support LimaCCDs Tango DS in taurus
Add an specialized device for LimaCCDs tango DS and support it via
the TangoFactory.registerDeviceClass mechanism

This Device class works around the impossibility to subscribe event in
image attributes in Tango 7.
This specialized device subscribes to 'image_count' attribute changes
and for each new image it pushes a new event with value-> image data
and source->  image attribute.