Add TangoSchemeTest device server
s device server is for unittest purposes.
It has a subgroup of attributes of the TangoTest device server.
This device server  allows to the developers play with its attributes,
e.g. modify value and properties, without interfere any other clients.

It is created using PyTango8 API, so it ONLY works with PyTango8.