Add tests for TangoAtribute
Add set of tests for write and read of Tango attributes, using the
TangoSchemeTest DS.

Note: In its current situation, all tests pass except 2, where
wvalue for read-only non scalar string attributes are expected to
be None and are read as empty tuples. This requires further
investigation to check if the behaviour is expected. TODO's are left
in place.