Remove {Taurus,Tango}StateAttribute classes
TaurusStateAttribute is just the same as TaurusAttribute except
that it implements a .isInformDeviceOfErrors() method which is
*deprecated*. TangoAttribute also implements this method.
TangoStateAttribute inherits from both TangoAttribute and
TaurusStateAttribute. But because of the (probably buggy)
implementation of its multiple inheritance,
TangoStateAttribute  is, in practice, just identical to
TangoAttribute. Therefore we can safely use TangoAttribute
and get rid of TangoStateAttribute. Here we assume that no
3rd party code ever used TaurusStateAttribute or
TangoStateAttribute classes explicitly (pretty safe
assumption)