Move TaurusDevice methods to TangoDevice and refactor
Move the TaurusDevice.getValueObj and cleanUp methods to TangoDevice.
Refactor the uses of getValueObj method.
Implement a dummy getDevState in TangoDevice.
Change TangoDevice.decode method to private.
Change the deprecation message in some TangoDevice methods related
with the State.
Fix the behaviour of TangoDevice.getState and TangoDevice.getSWState
to be backward compatible.