Rename method getDevState to state
Rename TaurusDevice.getDevState to TaurusDevice.state, and
make it a (read-only) property. This way the device state
can be accessed using both dev.state and dev.state().
Adapt all uses and reimplementations of getDevState
(since getDevState was introduced in a previous commit that
has not been yet released, there is no need to provide
bck-compat).
Note: this will break bck-compat for users intending to
call the PyTango.DeviceProxy.state command from a
TangoDevice (they will get a TaurusDevState instead of a
PyTango.DevState), but this is indeed a desirable outcome
since it makes the code scheme-agnostic and it is in line
with all other tep14 bck-incompatible changes regarding
the device state. Furthermore, the user can still access
the Tango command explicitly using
TangoDevice.getDeviceProxy().state