Adapt the tangoAttribute  to the new properties
Label,  range , alarm and warning attributes are converted to properties
in TaurusAttribute. TangoAttribute needs to be adapted.

Refactor the existing methods to adapt them to the new properties.
Fix the implementation to update the tango database in the setter methods.

Remove duplicated methods of the base class (TaurusAttribute).
 e.g. getMinAlarm, getMaxAlarm, ...

Move to deprecated getMInValue and  getMaxValue cretae alternative
getMinRange and getMaxRange.