Adapt taurusvaluestable to TEP14
Remove the old style use of configuration object.

Modify the different classes to use Quantities instead of floats and
ints. Use the _PintValidator to validate the inputs of the table.

Fix some wrong behaviour in TaurusValuesIOTableModel class
related with inputs that are out of range and out of limits.

Adapt the different classes to avoid deprecation warnings.