Fix case sensitive bug in taurusplot (partial)
taurusplot assumes that all the given models are case insensitive and
converts them to lower, making that the case sensitive models do not
work.

Fix it, avoiding to change to lower the model of the case sensitive
schemes.

IMPORTANT: TaurusPlot still considers that model names are always case
insensitive (tango-centric). This needs to be adapted.