Avoid to expand the expression in the full name
Evaluation AttributeName Validator getNames method returns the fullname
and simplename with the expanded expression. This causes errores with the
simple-name and labels.

Moreover make that the evaluation factory does not create a new attribute
if previously an attribute with the same expanded evaluation but different
formula was created.

Fix it, avoiding to expand the expresion in the full name.