[PINT] Use empty string for dimensionless abbreviated format
Make pint use empty string when formatting dimensionless units
(instead of "dimensionless") if the abbreviation format spec ("~")
is used.

IMPORTANT NOTE:
This is a change in our local pint. Ideally it should be done
upstream (in the official pint repo, see Pint issue #300). For the
moment we do it in our local copy.