Expand fullname in getNames method
EvaluationAttributeNameValidator getNames method does not expand
the fullname of its refs. This can cause errors when the factory
looks up for existing attributes.

Make the eval full name use full names of its references, e.g.:

The full name of eval:{a/b/c/d}+1 is currently:

  eval://localhost/@DefaultEvaluator/{a/b/c/d}+1

but it should be:

  eval://localhost/@DefaultEvaluator/{tango://foo:1234/a/b/c/d}+1
  (asuming that foo:1234 is the default tango authority)

Note: If the full name for the refs is not computable, the eval
attribute won't be created.