Update the modelFragmentName before the attach method
SetNodelName gets the fragment after call the attach method, doing
the post attach acctions  without getting the modelFragmentName.
The current implementation requires the use of the modelObj
to get the modelFragmentName.

Move the affected code before the call of the attach method and get the
validator from the Factory instead of uses the modelObj.