Fix bug-223: change order of cleanUp atexit registration
A simple script like this produces segfault when it exits:

  import taurus
  attr = taurus.Attribute('sys/tg_test/1/short_scalar')

The reason is that the TangoAttribute.cleanUp is called after PyTango's
own cleanup (causing event callback being called after receiver has been
deleted) due to the order in which the atexit registrations of Taurus
and PyTango are executed. Fix by delegating the atexit registration of
the the cleanUp of the factories to the factories themselves.