Fix #224 Taurus4: regressions and bugs in TaurusValuesTable
TaurusValuesTable has suffered some regressions:
-  cells do not resize to fit contents
-  support for writing string type is buggy.

Also, there is a bug related with the units.
Units are checked on model setting, but not on each event.
They could change and the table will show wrong info.
Even after changing values, the magnitudes are updated but the units aren't.

Fix these bugs:
- The resize polity is reverted.
- Fix the bug with the strings.
- Fix the bug with the units.