Fix bug#140 Unable to write an uchar scalar TangoAttribute
TangoAttribute raises a TypeError: "an integer is required" when you
try to write in a DevUChar attribute.

PyTango expects an int value while Taurus is trying to write a str
value.

Fix it by modifying the encode and decode behaviour for DevUChar
TangoAttributes. Decode  the DevUChar values to 1-element strings
and encode to int. Decode the DevUchar tango type to DataType.Bytes