Allow arbitrary characters in eval references
The patch for #241 by cfalcon fixes said issue but exposes an existing
bug in the validation of EvaluationAttribute names: the references form
part of the expression and therefore must comply with the limitations
for the python expressions (i.e., cannot contain "?" or "#" or "@").
Avoid this limitation by mangling the refs before matching the pattern
and unmangling after that. Also provide some tests that check this.