Fix bug-249
addSafe and removeSafe methods of SafeEvaluator class do not modify
the _originalSafeDict attribute used in resetmoSafe to restore the
dict, so the modification of the  safe_dict is not permanent.

Add a flag, "permanent" by default False, in  those methods
to do also the modification in the  _originalSafeDict, when the
flag is True.