Fix bug-262: create lex/yacc tab files only at run time
Creation of jdraw_lextab.py and jdraw_yacctab.py in the installation
dir (and therefore distributing them) creates problems for packaging
arch-independent packages. Change the logic of jdraws's new_parser so
that these files are created in "~/.taurus" and therefore they only
need to be created at runtime (the first time they are needed) and not
during build/install. As a side efect, less garbage is created during
build.