Fix bug-300 TaurusTrend fails exporting data in multiples files
TaurusTrend Export data to ASCII menu option does not work if you
try to save the data in multiples files.

If the data is in the same X-axis, it create n files where n is the number of
the ploting curves with the same data (the data of all curves).
but if the curves are not in the X-axis, the content of each files
is a warning message.

"Unable to display because abscissas are different. Curves will be saved
each one in its own file"

Fix it.