Fix testsuite (test modules imports)
All autodiscovered test modules must be importable from the top level of the
project, which by default is start_dir. Autodiscovered test modules do not
include taurus module and it can cause conflict with tests of other libraries.

Explicitly set top_level_dir to the parent of taurus.__file__ in order to
include taurus in the test module names and avoid the problem.