Adapt makeman to setuptools
the makeman script cannot find the scripts since now they are
dynamically created by setuptools. Adapt makeman todo a temporary
installation to generate the scripts.

Note that this is still a very fragile solution. We should consider
writing static man pages or leaving their creation to the
packaging, or finding a more robust method for creation.