Fix bug-338: make LE getValue() return out-of-range quantities
In the case of numerical attributes, TaurusValuelineEdit.getValue()
returns None when the value does not validate according to the line
edit's validator.

This is ok if the text is not a valid quantity representation, but it
is not ok when the text is a correct quantity representation and the
non-validation is due to it being out-of-range.

Fix by always returning the quantity, and implementing an independent
method `isValidText` to distinguish text that does not validate.

IMPORTANT: this commit may break some logic depending on the previous
(incorrect) behaviour.