Fix bug-339: Avoid write when focus lost in TaurusValueSpinBox
TaurusValueSpinBox connects its edittingFinished signal to its
line edit's writeValue(). This allows to use the enter key for applying
values,but has the unintended effect of also writing values when focus
is lost (even when autoapply is not True).

Fix by delegating the key presses in TaurusValueSpinBox to its LineEdit
instead of connecting the edittingFinished signal to writeValue.
Note that the line edit already connects the edittingFinished to a
method that checks if autoapply is True before calling writeValue.