Add TEPs access in Taurus' documentation
TEPs are moved from SF wiki to GitHub. This is annoying situation since
the reference of old and new TEPs is different,
in order to normalize the access point to the TEP, Taurus'
documentation has to point to the place where the TEP are hosted
without distintion and transparently for the users.

Add a link in the main documentation menu to the TEP index web page,
and implement a redirect system to access to the TEPs via Taurus
web page with the following syntax:

www.taurus-scada.org/tep/?TEPXX.EXT

where:
- XX is the number of the TEP. e.g. 1,2,...
- EXT is the extension of the TEP document. e.g. md,...