Fix-365: Refactor TangoAttribute push_event
Taurus4 introduced a regression related to the merge of TaurusAttribute
and TaurusConfiguration. The TangoAttribute.push_event method handles
all events but it doesn't do it properly for configuration-related
error events received before the TangoAttribute has been initialized.

In this situation, the TangoAttribute ends up flagged as subscribed
but the subscription was not actually done, preventing to become
properly subscribed when a non-error event comes later on.
It also may incorrectly deactivate the polling mechanism.

Refactor the push_event code to clean it and correctly handle conf
errors.