Fix #367: Revert to uchar as uint8 interpretation
Taurus4 introduced a change in the interpretation of Tango uchar values.
In Taurus3 uchar values were interpreted as integers (for arrays,
uint8). In Taurus4 the interpretation was changed to
treat uchar as one-character strings (and for arrays, dtype='S1').

Revert to the pre-Taurus4 interpretation since this actually changes
the expected behaviour.