Fix drawback in Taurusgui --tango-host
Taurusgui --tango-host=$TANGO_HOST raises a TaurusException:
"Invalid default Tango authority name" with taurus4. According
the documentation it should be work, and it is a drawback introduced
in taurus4.

This happens due to the change of taurus4 to forze to use strict URIs.
Fix it, translating the given HOST_NAME to a taurus authority URI.