Deprecate TaurusGui.createConsole and the "CONSOLE" keyword
All TaurusGuis attempt to add a TaurusConsole panel by default.
Deprecate this behaviour, as well as the associated createConsole()
method and the "CONSOLE" configuration keyword in favour of treating
TaurusConsole just as any other panel that can be added using the
regular TaurusGUI panel API.

In compensation, add the TaurusConsole widget to the default catalog for
the "new panel" action.