Fix #361 Taurus widget shows blanck label
Taurus widget shows blanck label instead of the attribute
representation when it has defined an invalid units
(for pint library).

An UnicodeDecodeError is raising when pint try to decode
the unit. Taurus should retrun dimensionless unit in these cases.

Fix it, handling the exception and return dimensionless unit when
taurus could not get a valid unit.