Add class support in eval devices
Eval devices documentation states that you can provide a
SafeEvaluator - derived class to be used as a custom evaluator,
but this is not actually correct (it was like that for Taurus3,
but not anymore in Taurus 4, where the class actually needs to
inherit also from TaurusDevice).

Change the implementation so that when using "eval:@foo.C" the C
class no longer needs to inherit from anything and instead the
resulting device object is a EvaluationDevice with all symbols
from C added to its safe dict.

Additionally, support a new syntax: "eval:c=foo.C()" ("c=" is
actually optional) similar to the previous one, but an instance
of C is created and the symbols added to the safe Dict are the
instance's and not those of the class (this allows for persistence
of instances and facilitates implementing writing support in eval).