Refacto EvalFactory.getDevice
- Simplify the implementation and use the dir() method instead of __dict__,
which is more robust.

- Do not add all symbols from instance (just add instance) to the safe
dict. This is needed because if a user attempts to access a "property" from
the safedict, she would get the return from the property at the moment of
creating the safedict instead of calling the getter behind the scenes. With
this change, the user is forced to use the instance symbol and thus would
actually use the getter.