Implement TaurusBaseComponent Formatting API
As part of the solution of bug  # 392 an API is needed to allow
to modify the represented format of the Taurus attributes in
the different widgets (in other words modify its View)

Create an API to manage the `format` in TaurusBaseComponent class.
The API allows to change the formatting of the represented values via
class members or via methods.

Add `FORMAT` and `defaultFormatDict` class members:

- `FORMAT` accept a python string formatting [1] or a callable that returns
a python string formatting.

e.g.
 # with string formatting
 TaurusBaseComponent.FORMAT = "{:.2e}"
 or
 # with callable
 def baseFormatter(dtype=None, basecomponent=None, **kwargs):
     return "{:.1f}"

 TaurusBaseComponent.FORMAT = baseFormatter

`defaultFormatDict` is the dictionary with the codification between
attribute type and the string formmatting. It can be extended or modified.

e.g.
 # Add a custom string formmating for str types: e.g. using its repr()
 TaurusBaseComponent.defaultFormatDict.update({"str": "{!r}"})

Add `setFormat` and `resetFormat` methods to manage the `FORMAT`
at the instance level. Other internal methods are implemented.

The displayValue() method is modified to use the string formatting.

Implement a default formatter based on the use of `precision`.

[1] https://docs.python.org/2/library/string.html