Allow subscribing to events w/ or w/o first event
Emitting first event during the subscription process with the most recent
value may not be always desired.

This is the case of the TangoAttributeEG class in Sardana, a subclass of
the EventGenerator. It may work with the attribute which serves as a data
pipe for transferring scan results. It is not desired to receive an event
with the data corresponding to the previous scan while subscribing to events
of the new scan.

Allow control whether to emit the first event during the subscription process.