Added legend option and ModelChooser
Added option to widget menu for show and hide the legend from
the plot widget. This legend will be updated automatically with
the curves added or removed from the widget.
Also added to widget menu the calling for a modelChooserTool
that allows choose a model for X and Y axis.