Behaviour of Y2ViewBox class in CurveProperties.
The TaurusPlot (widget) is the manager to create a Y2ViewBox and
specify to a CurvePropertiesTool that can be used in the
CurveAppearanceChooserDialog for swapping curves between main
and Y2 axis.
The TaurusPlot widget now is a bridge between the Y2ViewBox and
the dialog.
If the dialog doesn't recive any Y2ViewBox object, the groupBox
that manages the Y1 and Y2 axis, will be disabled.