Add config property in Y2ViewBox class
Add config property in Y2ViewBox class and delegate in TaurusPlot.
Now, the Y2ViewBox can remember the curves associated to him self
and the widget(TaurusPlot) can ask him for add each curve in
the correct ViewBox (Axis in this case).
Y2ViewBox has the property for save their curves, but is the
widget, which has to decide how add this curves, having the
info about the Y2 axis curves.