Remove labelText property from TaurusValue
PR #496 introduced a new assumption that a label widget of a TaurusValue
must be a TaurusLabel or its subclass. But custom widgets were also
accepted in the past e.g. Sardana PMTV. Change it since this compromises
the API too much and remove a way of setting a label text using the
TaurusValue. Instead a label text should be set directly on the label
widget.