Fix issues with global setting of default format
Using --default-format is not working properly due to uninitialized
stuff when setFormat is called. Also, if used, the customization of
FORMAT at a class level is lost. Fix this by ensuring the required
variables are initialized before calling setFormat in the
initialization of TaurusBaseComponent, and doing it *only*
if FORMAT is not set at class level.
Also, change TaurusValue to delegate onSetFormatter to the read widget,
instead of delegating getFormat and setFormat