Fix Taurus may emit events with outdated data
In the Tango scheme there are three occassions that that may
update the attribute ( __attr_value, __attr_err).

These methods are:

- read (executed in the main thread)
- push_event (executed in the Tango event consumer thread)
- poll (executed in the Taurus polling thread)

Current implementation may lead to emitting events with the
outdated values. The polling mechanism reads the Tango
attributes asynchronously (first drop the read requests
and later on collect the replies). The more up to date event
value may get overriden by an older value corresponding to
the polling request dropped prior to the event reception.

Add a RLock to protect the critical section (update the attribute).

Fix #216