Fix #612: ignore extras info when checking requirements
Under some circumstances (e.g. when installing taurus with pip from
a tar.gz created with the sdist command), the
`pkg_resources.Requirement` objects returned when interrogating taurus
for its requirements include `.marker` information, whereas in other it
does not. This mreaks the logic of `taurus.check_requirements()`.
As a quick workaround, remove the markers info from the requirement
string.