Fix bug in setting serialization mode
TaurusModel's logic for setting the serialization mode is wrong due to
a tabulation problem. This leads to TaurusDevices always being
initialized with serialization mode 'Serial', ignoring the mode set
in the Manager (by default, 'Concurrent'). Fix the tabulation issue and
restore the intended logic.

Warning: be aware that because the Concurrent mode was in practice not
set until now (in spite of being the intended situation), some code may
not have been properly tested under Concurrent serialization mode and
therefore potentially bugs which have been hidden may now suface.