Fix #707 (stepping in line edits and spin boxes does not work)
The step up/down feature of TaurusValueLineEdit (e.g. using
the up/down keys or the arrows of a TaurusValueSpinBox) was broken when
merging #669 which implemented fragments support in
TaurusValueLineEdit for allowing the display of the wvalue without units
Now we see that supporting fragments generically in TaurusValueLineEdit
is problematic (e.g. it makes little sense to use "rvalue" as a
fragment since "wvalue" will be modified when eventually applying).
Therefore for now limit the fragment support in line edits
to the "wvalue.magnitude" case.

Also fix a hook error introduced in some previous merge.