Make TaurusSpinBox delegate step size to line edit
TaurusSpinBox composes a TaurusValueLineEdit and delegates many of
its features to it. But the "singleStep" property is not delegated and
this causes inconsistent handling of step increments (e.g. pressing the
keyboard up/down arrows ignores the step size, while using the up/down
arrows of the spinbox widget takes it into account). Fix this by
delegating the "singleStep" property accessors too.