Improve consistency in stepping for taurus spinbox and line edit
- The TaurusValueLineEdit and TaurusSpinbox use the CTRL key modifier
both for multiplying the singleStep size and for "force-applying".
Avoid this conflict by removing the "multiplier" role of the CTRL key.
- Similarly, The ALT key modifier (used for dividing the singleStep) is
also problematic when using it with the mouse as it conflicts with "drag
window" actions. Remove that "divider" role as well.
- Refactor the TaurusSpinBox.stepBy implementation to delegate to the
composed line edit in order to offer more consistency.
- Also delegate handling the wheel events for consistency.