Fix #757 (ReferenceError in TangoAttrValue)
TangoAttrValue keeps a weak reference to its associated TangoAttribute,
which is used for accessing the "configuration" API. But, unless the
TangoAttribute is hard-referenced elsewhere, the weak ref would be
broken. This problem has been hidden due to issue #511, but it is
evident now since we fixed it in #692.
Fix #757 by recreating the attribute when needed.