Use the original for-loop for supporting index values
The original implementation of getDisplayValue used a for-loop
for supporting the modelIndex feature which was changed in the previous 
commit in an attempt to make it more efficient. But this new implementation
fails on 2D attributes, as pointed out by @CorF4. Recover the original 
implementation but keep the try-except protection.