Make external.qt update os.environ[QT_API]
If QT_API is not declared in the environment, the choice of binding
done by taurus.external.qt may sometimes differ from that of some of the
modules used as dependencies (e.g. guidata or spyder) causing conflicts
of imported bindings. Work this around by defining QT_API in the
os.environ (this is a temporary change affecting only the current
process).