Revert "Set pyqt5 as the default qt binding"
Go back to setting PyQt4 as the default binding if one is not
already loaded or selected via QT_API environment variable.
The reason is that the plot (qwt5) functionality is not yet
fully supplied by the taurus_pyqtgraph plugin and therefore
it is safer not to do this change yet to unaware users.

This reverts commit 08d61e2cb4.